+++ packages/esm-patient-common-lib/src/orders/types/test-order.ts
@@ -1,13 +1,10 @@
 import type { OrderBasketItem } from './order';
 
 export interface TestOrderBasketItem extends OrderBasketItem {
-  testType?: {
+  testType: {
     label: string;
     conceptUuid: string;
   };
-  instructions?: string;
-  previousOrder?: string;
   orderReason?: string;
-  orderNumber?: string;
   specimenSource?: string;
 }
