+++ packages/esm-patient-orders-app/src/components/test-order.component.tsx
@@ -30,7 +30,7 @@ const TestOrder: React.FC<TestOrderProps> = ({ testOrder }) => {
   const tableHeaders: Array<{ key: string; header: string }> = [
     {
       key: 'testType',
-      header: testOrder?.orderType?.display,
+      header: testOrder.orderType.display,
     },
     {
       key: 'result',
