+++ packages/esm-patient-orders-app/src/order-basket/order-basket.workspace.tsx
@@ -121,11 +121,11 @@ const OrderBasket: React.FC<DefaultPatientWorkspaceProps> = ({
             name="order-basket-slot"
           />
           {config?.orderTypes?.length > 0 &&
-            config?.orderTypes?.map((orderType) => (
+            config.orderTypes.map((orderType) => (
               <GenericOrderType
                 key={orderType.orderTypeUuid}
                 orderTypeUuid={orderType.orderTypeUuid}
-                orderableConcepts={orderType.orderableConcepts}
+                orderableConceptSets={orderType.orderableConceptSets}
                 closeWorkspace={closeWorkspace}
               />
             ))}
