+++ packages/esm-patient-orders-app/src/order-basket/order-types/generic-order-panel.scss
@@ -76,6 +76,6 @@
 
 .chevron {
   svg {
-    fill: black;
+    fill: currentColor;
   }
 }
