+++ packages/esm-patient-orders-app/translations/en.json
@@ -88,8 +88,8 @@
   "saveOrder": "Save order",
   "saving": "Saving",
   "searchAgain": "search again",
-  "searchConcepts": "Search concepts",
   "searchFieldOrder": "Search for {{orderType}} order",
+  "searchOrderables": "Search orderables",
   "searchResultsMatchesForTerm_one": "{{count}} results for \"{{searchTerm}}\"",
   "searchResultsMatchesForTerm_other": "{{count}} results for \"{{searchTerm}}\"",
   "searchTable": "Search table",
